Proposed package layout
