+++ .gitignore
@@ -1,4 +1,4 @@
-
-*.pyc
+__pycache__/
+*.py[cod]
 
 test.*.py
